fix: remove node-types duplicate codes
