+++ packages/otelbin/src/components/react-flow/node-types/ProcessorsNode.tsx
@@ -1,105 +1,23 @@
 // SPDX-FileCopyrightText: 2023 Dash0 Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-import React, { memo, useState } from "react";
-import { Handle, Position } from "reactflow";
-import { useEditorRef, useFocus } from "~/contexts/EditorContext";
-import { FlowClick } from "../FlowClick";
 import type { IData } from "../FlowClick";
+import Node from "./Node";
 import ProcessorIcon from "~/components/assets/svg/processor.svg";
-import { iconColor } from "./ExportersNode";
+import { handleStyle } from "./Node";
+import { memo } from "react";
+import { Handle, Position } from "reactflow";
 
 const ProcessorsNode = ({ data }: { data: IData }) => {
-	const [hovered, setHovered] = useState(false);
-	const editorRef = useEditorRef();
-	const { isFocused } = useFocus();
-
-	const customNodeHeaderStyle = {
-		borderRadius: "8px 8px 0px 0px",
-	};
-
-	const customNodeStyles = {
-		display: "flex",
-		alignItems: "center",
-		justifyContent: "center",
-		width: "100%",
-		height: "65%",
-		background: hovered ? "#40454E" : "#30353D",
-		transition: "background-color 0.3s ease-in-out",
-		borderRadius: "0px 0px 8px 8px",
-		paddingBottom: "6px",
-		paddingTop: "6px",
-		borderLeft: `1px solid ${hovered ? "#6D737D" : "#40454E"}`,
-		borderRight: `1px solid ${hovered ? "#6D737D" : "#40454E"}`,
-		borderBottom: `1px solid ${hovered ? "#6D737D" : "#40454E"}`,
-	};
-
-	function handleClickNode(event: React.MouseEvent<HTMLDivElement, MouseEvent>) {
-		FlowClick(event, data, editorRef);
-	}
-
-	const label = data.label || "";
-
-	const splitLabel = typeof label === "string" ? label.split("/") : [];
-	const hasSlash = splitLabel.length > 1;
 	return (
-		<>
-			<div
-				className={`flex h-20 max-h-[81px] max-w-[121px] w-[120px] flex-col items-center rounded-lg shadow-node mx-3 ${
-					isFocused === data.id ? "animate-processorFocus" : ""
-				}`}
-			>
-				<div
-					style={customNodeHeaderStyle}
-					className="px-3 bg-blue-500 text-center text-xs font-medium h-[35%] overflow-hidden whitespace-nowrap overflow-ellipsis w-full flex items-center justify-center"
-				>
-					{splitLabel[0]}
-				</div>
-				<div
-					style={customNodeStyles}
-					className="cursor-pointer flex-col"
-					onClick={handleClickNode}
-					onMouseEnter={() => setHovered(true)}
-					onMouseLeave={() => setHovered(false)}
-				>
-					<Handle
-						type="target"
-						position={Position.Left}
-						style={{
-							backgroundColor: "rgb(44 48 70 / 0%)",
-							borderColor: "rgb(44 48 70 / 0%)",
-						}}
-					/>
-					<div
-						className={`flex w-full flex-col items-center justify-center gap-y-1 px-2 ${
-							splitLabel[1] && splitLabel[1].length > 0 && "mt-[2px]"
-						}`}
-					>
-						<div style={iconColor(hovered)}>
-							<ProcessorIcon />
-						</div>
-						{hasSlash && (
-							<div
-								className={`${
-									hovered ? "text-neutral-900" : "text-neutral-600"
-								} text-[10px] font-normal  overflow-hidden whitespace-nowrap overflow-ellipsis max-w-[90%]`}
-							>
-								{splitLabel[1]}
-							</div>
-						)}
-					</div>
-					<Handle
-						type="source"
-						position={Position.Right}
-						id="processor-a"
-						style={{
-							backgroundColor: "rgb(44 48 70 / 0%)",
-							borderColor: "rgb(44 48 70 / 0%)",
-						}}
-					/>
-				</div>
-			</div>
-		</>
+		<Node
+			icon={<ProcessorIcon />}
+			data={data}
+			type="processor"
+			handle1={<Handle type="target" position={Position.Left} style={handleStyle} />}
+			handle2={<Handle type="source" position={Position.Right} id="processor-a" style={handleStyle} />}
+		/>
 	);
 };
+
 export default memo(ProcessorsNode);
